feat(TMRX-1557): refactor content bucket 3
