+++ packages/ts-newskit/src/slices/content-bucket-3/styles.tsx
@@ -1,0 +1,9 @@
+import { Block, getMediaQueryFromTheme, styled } from 'newskit';
+
+export const LeadArticles = styled(Block)`
+  ${getMediaQueryFromTheme('md')} {
+    hr[aria-label='article-divider-horizontal'] {
+      display: none;
+    }
+  }
+`;
